# openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "RealAdvisor Inbox API",
		"version": "1.0.0",
...
	},
	"servers": [
		{
			"url": "/",
			"description": "Current inbox instance"
		}
	],
	"security": [
		{
			"AccessSession": []
		}
	],
...
	"paths": {
...
		"/api/v1/config": {
			"get": {
				"operationId": "getGetConfiguration",
				"summary": "Get configuration",
				"tags": ["System"],
				"description": "",
				"parameters": [],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"domains": {
											"type": "array",
											"items": {
												"type": "string"
											}
										},
										"emailAddresses": {
											"type": "array",
											"items": {
												"type": "string"
											}
										},
										"mode": {
											"type": "string",
											"enum": ["live", "synthetic"]
										},
										"canCreateMailboxes": {
											"type": "boolean"
										},
										"canDeleteMailboxes": {
											"type": "boolean"
										},
										"classifierPreview": {
											"type": "boolean"
										},
										"classifiersEnabled": {
											"type": "boolean"
										},
										"canManageWebhooks": {
											"type": "boolean"
										},
										"canManageClassifiers": {
											"type": "boolean"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid input",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid authentication; the Access edge may redirect anonymous browsers to login."
					},
					"403": {
						"description": "Operation denied",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"413": {
						"description": "Request exceeds 256,000 bytes",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"415": {
						"description": "Use application/json",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"security": [
					{
						"AccessSession": []
					},
					{
						"InboxApiKey": []
					}
				]
			}
		},
		"/api/v1/mailboxes": {
			"get": {
				"operationId": "getListMailboxes",
				"summary": "List mailboxes",
				"tags": ["Mailboxes"],
				"description": "",
				"parameters": [],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Mailbox"
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid input",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid authentication; the Access edge may redirect anonymous browsers to login."
					},
					"403": {
						"description": "Operation denied",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"413": {
						"description": "Request exceeds 256,000 bytes",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"415": {
						"description": "Use application/json",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"security": [
					{
						"AccessSession": []
					},
					{
						"InboxApiKey": []
					}
				]
			},
			"post": {
				"operationId": "postCreateMailbox",
				"summary": "Create mailbox",
				"tags": ["Mailboxes"],
				"description": "In live mode, only users listed in MAILBOX_ADMINS can create mailboxes, and creation must be enabled. The email must use ingest.realadvisor.com, with a 1\u201364 character lowercase local part containing letters, numbers, dots, underscores or hyphens, starting and ending with a letter or number; consecutive dots are forbidden. Case is normalized to lowercase. Registration and system folders are committed atomically. Existing addresses return 409. The shared routing rule immediately permits this registered inbox, and sending uses the matching @realadvisor.com identity. Google Workspace/Groups forwarding for the public address is a separate step. Local synthetic mode remains unrestricted.",
				"parameters": [],
				"responses": {
					"201": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Mailbox"
								}
							}
						}
					},
					"400": {
						"description": "Invalid input",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid authentication; the Access edge may redirect anonymous browsers to login."
					},
					"403": {
						"description": "Operation denied",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"413": {
						"description": "Request exceeds 256,000 bytes",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"415": {
						"description": "Use application/json",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"format": "email",
										"maxLength": 254
									},
									"name": {
										"type": "string",
										"minLength": 1,
										"maxLength": 120
									}
								},
								"required": ["email", "name"]
							}
						}
					}
				}
			}
		},
		"/api/v1/mailboxes/{mailboxId}": {
			"get": {
				"operationId": "getGetMailbox",
				"summary": "Get mailbox",
				"tags": ["Mailboxes"],
				"description": "",
				"parameters": [
					{
						"$ref": "#/components/parameters/mailboxId"
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Mailbox"
								}
							}
						}
					},
					"400": {
						"description": "Invalid input",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid authentication; the Access edge may redirect anonymous browsers to login."
					},
					"403": {
						"description": "Operation denied",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"413": {
						"description": "Request exceeds 256,000 bytes",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"415": {
						"description": "Use application/json",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"security": [
					{
						"AccessSession": []
					},
					{
						"InboxApiKey": []
					}
				]
			},
			"put": {
				"operationId": "putUpdateMailboxSettings",
				"summary": "Update mailbox settings",
				"tags": ["Mailboxes"],
				"description": "Stores the display setting; does not change the fixed sender email address.",
				"parameters": [
					{
						"$ref": "#/components/parameters/mailboxId"
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Mailbox"
								}
							}
						}
					},
					"400": {
						"description": "Invalid input",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid authentication; the Access edge may redirect anonymous browsers to login."
					},
					"403": {
						"description": "Operation denied",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"413": {
						"description": "Request exceeds 256,000 bytes",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"415": {
						"description": "Use application/json",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"settings": {
										"type": "object",
										"properties": {
											"fromName": {
												"type": "string",
												"maxLength": 120
											}
										},
										"required": ["fromName"]
									}
								},
								"required": ["settings"]
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "deleteDeleteSyntheticMailbox",
				"summary": "Delete synthetic mailbox",
				"tags": ["Mailboxes"],
				"description": "Local synthetic mode only. Live deployments return 403.",
				"parameters": [
					{
						"$ref": "#/components/parameters/mailboxId"
					}
				],
				"responses": {
					"204": {
						"description": "Success; no response body"
					},
					"400": {
						"description": "Invalid input",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid authentication; the Access edge may redirect anonymous browsers to login."
					},
					"403": {
						"description": "Operation denied",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"413": {
						"description": "Request exceeds 256,000 bytes",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"415": {
						"description": "Use application/json",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"500": {
						"description": "Internal error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
...
		"/api/v1/tags": {
			"get": {
				"operationId": "listTags",
				"summary": "List shared tags",
				"tags": ["Tags"],
				"description": "Shared catalog across mailboxes, available to all authenticated inbox users.",
				"parameters": [],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Tag"
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid input or confirmation missing",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"401": {
						"description": "Access login required",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"403": {
						"description": "Operation denied",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"404": {
						"description": "Tag or conversation not found in this mailbox",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"409": {
						"description": "Duplicate name or conflicting reference",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"security": [
					{
						"AccessSession": []
					},
					{
						"InboxApiKey": []
					}
				]
			},
			"post": {
				"operationId": "createTag",
				"summary": "Create shared tag",
				"tags": ["Tags"],
				"description": "Names are trimmed and case-insensitively unique.",
				"parameters": [],
				"responses": {
					"201": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tag"
								}
							}
						}
					},
					"400": {
						"description": "Invalid input or confirmation missing",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"401": {
						"description": "Access login required",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"403": {
						"description": "Operation denied",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"404": {
						"description": "Tag or conversation not found in this mailbox",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"409": {
						"description": "Duplicate name or conflicting reference",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TagInput"
							}
						}
					}
				}
			}
		},
		"/api/v1/tags/{tagId}": {
			"put": {
				"operationId": "updateTag",
				"summary": "Update tag",
				"tags": ["Tags"],
				"description": "Edit tag name and color.\n\nStable UUID references preserve every assignment across renames.",
				"parameters": [
					{
						"name": "tagId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tag"
								}
							}
						}
					},
					"400": {
						"description": "Invalid input or confirmation missing",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"401": {
						"description": "Access login required",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"403": {
						"description": "Operation denied",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"404": {
						"description": "Tag or conversation not found in this mailbox",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"409": {
						"description": "Duplicate name or conflicting reference",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TagInput"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "deleteTag",
				"summary": "Delete tag",
				"tags": ["Tags"],
				"description": "Delete shared tag and all assignments.\n\nConfirm with the user first. Permanently deletes the tag and all assignments, including removal records, across every mailbox.",
				"parameters": [
					{
						"name": "tagId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "confirm",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"enum": ["true"]
						}
					}
				],
				"responses": {
					"204": {
						"description": "Success"
					},
					"400": {
						"description": "Invalid input or confirmation missing",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"401": {
						"description": "Access login required",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"403": {
						"description": "Operation denied",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"404": {
						"description": "Tag or conversation not found in this mailbox",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"409": {
						"description": "Duplicate name or conflicting reference",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
...
		"/api/v1/tag-groups": {
			"get": {
				"tags": ["Tags"],
				"summary": "List shared tag groups",
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/TagGroup"
									}
								}
							}
						}
					}
				},
				"security": [
					{
						"AccessSession": []
					},
					{
						"InboxApiKey": []
					}
				]
			},
			"post": {
				"tags": ["Tags"],
				"summary": "Create tag group",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TagGroupInput"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TagGroup"
								}
							}
						}
					}
				},
				"description": "Create a tag group and managed Jev classifiers (administrator only)."
			}
		},
		"/api/v1/tag-groups/{id}": {
			"put": {
				"tags": ["Tags"],
				"summary": "Update tag group",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TagGroupInput"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TagGroup"
								}
							}
						}
					},
					"403": {
						"description": "Administrator access required"
					},
					"409": {
						"description": "Revision conflict or multiple existing selections"
					}
				},
				"description": "Update a group, tags and its shared Jev instructions (administrator only)."
			},
			"delete": {
				"summary": "Delete tag group",
				"operationId": "deleteTagGroup",
				"tags": ["Tags"],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"in": "query",
						"name": "confirm",
						"required": true,
						"schema": {
							"type": "string",
							"enum": ["true"]
						}
					},
					{
						"in": "query",
						"name": "revision",
						"required": true,
						"schema": {
							"type": "integer",
							"minimum": 1
						}
					}
				],
				"responses": {
					"204": {
						"description": "Group deleted; tags retired and assignments removed"
					},
					"400": {
						"description": "Invalid or unconfirmed request"
					},
					"403": {
						"description": "Administrator required"
					},
					"404": {
						"description": "Group not found"
					},
					"409": {
						"description": "Group revision changed"
					}
				},
				"description": "Delete a tag group and retire its tags and classifiers, preserving historical runs."
			}
		},
...
		"/api/v1/api-keys/{id}": {
			"delete": {
				"tags": ["API keys"],
				"summary": "Revoke API key",
				"description": "Revoke an API key (administrator only).\n\nNew requests using the key are denied immediately. Already accepted requests may finish. Existing webhook subscriptions remain active and administrators retain control.",
				"security": [
					{
						"AccessSession": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Revoked"
					},
					"403": {
						"description": "Administrator access required"
					},
					"404": {
						"description": "Not found"
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"Error": {
				"type": "object",
				"properties": {
					"error": {
						"type": "string"
					},
					"details": {
						"type": "string"
					}
				},
				"required": ["error"]
			},
			"Mailbox": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"name": {
						"type": "string"
					},
					"settings": {
						"type": "object",
						"properties": {
							"fromName": {
								"type": "string"
							}
						}
					},
					"created_by": {
						"type": "string",
						"nullable": true
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": ["id", "email", "name", "settings"]
			},
...
			"TagInput": {
				"type": "object",
				"additionalProperties": false,
				"required": ["name", "color"],
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 80
					},
					"color": {
						"type": "string",
						"pattern": "^#[0-9a-fA-F]{6}$",
						"example": "#2563eb"
					}
				}
			},
...
		},
		"parameters": {
			"mailboxId": {
				"name": "mailboxId",
				"in": "path",
				"required": true,
				"schema": {
					"type": "string"
				},
				"example": "privacy@ingest.realadvisor.com",
				"description": "Mailbox address. Live mailboxes: privacy@ingest.realadvisor.com and info@ingest.realadvisor.com. URL-encode path values."
			},
...
		},
		"securitySchemes": {
			"AccessSession": {
				"type": "apiKey",
				"in": "cookie",
				"name": "CF_Authorization",
				"description": "Verified Cloudflare Access browser session; human membership and roles apply. Do not copy session cookies to integrations."
			},
			"InboxApiKey": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "inbox_<64 hex characters>",
				"description": "Created by an administrator in Settings \u2192 API keys. Scope and mailbox restrictions apply; revoked or expired keys return 401."
			}
		}
	}
}